Fixed issue with few results being returned by elastic search.
Previously using the normalized monarch id as the doc id for each variable, which means it gets overwritten for each version.
For example if you have the same variable in 200 datasets we were only getting 1 result because on the last one would be there.
Created a new id that combines the dbgap variaible_id + dbgap study_id + normalized id