Chore: Improving unit tests for the new objectified dug stuff
This change starts building out unit tests for the new objectified dug
code. It also make some configuration and structural changes to clean up
and facilitate the testing. Particularly of note, the `bin/dug pytests`
script will no longer creates a new virtualenv, instead it will simply
run in the calling environment (this was doing a lot of
unnecessary work for users with PyEnv)